refactor(guide): 공식 가이드 순위 섹션을 OfficialRankingSection 으로 분리
CRITERIA + LeaderboardMock + RankingSection 을 프레젠테이션 컴포넌트로 이동.
마크업/클래스 동일 — 시각 불변. 미사용 아이콘 6종 import 정리. 604→317 LOC (B).
